pyproject.toml
[tool.poetry]
name = "Fastapi Boilerplate"
version = "0.0.1"
description = "A boilerplate for fastapi projects"
authors = ["Ian Pirro <ianpirro@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api" },
    { include = "application" },
    { include = "domain" },
    { include = "infrastructure" },
    { include = "persistence" },
]

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.6.0"
anyio = "3.7.1"
attrs = "21.4.0"
bcrypt = "4.0.1"
certifi = "2023.7.22"
cffi = "1.16.0"
click = "8.1.7"
dnspython = "2.4.2"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
exceptiongroup = "1.1.3"
execnet = "1.9.0"
fastapi = "0.104.1"
greenlet = "3.0.1"
h11 = "0.14.0"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.25.1"
idna = "3.4"
iniconfig = "1.1.1"
isort = "5.12.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.3"
orjson = "3.9.10"
passlib = "1.7.4"
pluggy = "1.0.0"
psycopg2 = "2.9.9"
py = "1.11.0"
pyasn1 = "0.5.0"
pycparser = "2.21"
pydantic = "2.5.1"
pydantic-extra-types = "2.1.0"
pydantic-settings = "2.1.0"
pydantic-core = "2.14.3"
pyparsing = "3.0.9"
pytest = "7.1.2"
pytest-cache = "1.0"
pytest-subtests = "0.8.0"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
starlette = "0.27.0"
tomli = "2.0.1"
typing-extensions = "4.8.0"
ujson = "5.8.0"
uvicorn = "0.24.0.post1"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
requests = "^2.31.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
setuptools = "^69.0.2"
alembic = "^1.13.0"
packaging = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
